backend/comprenanto_backend_structured/requirements.txt
fastapi
uvicorn
python-jose
python-multipart
python-dotenv